refactor(rest): extract admin audit-perf/detail routes into routes/admin_audit_perf.py
Moves the 21 remaining admin /audit/* routes out of create_app's closure
into routes/admin_audit_perf.py: issue-target config (GET/PUT), manual
issue-bot sync (per-run and per-row), the perf dashboard (perf,
perf/thresholds, perf/workers, perf/hotspots), the frontend-load and
render snapshots + hotspots, the per-audit-row detail downloads
(source/profile/log/metrics-history/client-metrics/profile/stats),
metrics cleanup, the audit summary, and the audit log list.

load_issue_target_config, run_issue_bot_for and
run_issue_bot_for_conversion move with them as plain (pool, dict)
functions — no RestContext needed. create_app keeps
_run_issue_bot_for/_run_issue_bot_for_conversion bound to them for the
issue-bot poller, which still lives in the closure.
